lidar_test/requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
attrs==23.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.108.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.0.4
numpy==2.0.1
opencv-python==4.10.0.84
outcome==1.3.0.post0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
pyserial==3.5
PySocks==1.7.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.15.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.32.0.post1
trio==0.23.1
trio-websocket==0.11.1
twilio==8.11.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.25.0
wsproto==1.2.0
yarl==1.9.4
